Fix packages; LDFLAGS
