+++ pkg/creds/types.go
@@ -17,22 +17,14 @@
 package creds
 
 import (
-	"strconv"
 	"sync"
-	"time"
 
+	"github.com/netflix/weep/pkg/aws"
 	"github.com/netflix/weep/pkg/metadata"
+	"github.com/netflix/weep/pkg/types"
 
 	"github.com/aws/aws-sdk-go/aws/credentials"
 )
-
-type AwsCredentials struct {
-	AccessKeyId     string `json:"AccessKeyId"`
-	SecretAccessKey string `json:"SecretAccessKey"`
-	SessionToken    string `json:"SessionToken"`
-	Expiration      Time   `json:"Expiration"`
-	RoleArn         string `json:"RoleArn"`
-}
 
 type RefreshableProvider struct {
 	sync.RWMutex
@@ -40,8 +32,8 @@ type RefreshableProvider struct {
 	client        HTTPClient
 	retries       int
 	retryDelay    int
-	Expiration    Time
-	LastRefreshed Time
+	Expiration    types.Time
+	LastRefreshed types.Time
 	Region        string
 	RoleName      string
 	RoleArn       string
@@ -58,7 +50,7 @@ type CredentialProcess struct {
 }
 
 type ConsolemeCredentialResponseType struct {
-	Credentials *AwsCredentials `json:"Credentials"`
+	Credentials *aws.Credentials `json:"Credentials"`
 }
 
 type ConsolemeCredentialRequestType struct {
@@ -78,61 +70,12 @@ type ConsolemeCredentialErrorMessageType struct {
 	RequestID     string `json:"request_id"`
 }
 
-type Time time.Time
-
-// MarshalJSON is used to convert the timestamp to JSON
-func (t Time) MarshalJSON() ([]byte, error) {
-	return []byte(strconv.FormatInt(Time(t).Unix(), 10)), nil
-}
-
-// UnmarshalJSON is used to convert the timestamp from JSON
-func (t *Time) UnmarshalJSON(s []byte) (err error) {
-	r := string(s)
-	q, err := strconv.ParseInt(r, 10, 64)
-	if err != nil {
-		return err
-	}
-	*(*time.Time)(t) = time.Unix(q, 0)
-	return nil
-}
-
-// Add returns t with the provided duration added to it.
-func (t Time) Add(d time.Duration) time.Time {
-	return time.Time(t).Add(d)
-}
-
-// Unix returns t as a Unix time, the number of seconds elapsed
-// since January 1, 1970 UTC. The result does not depend on the
-// location associated with t.
-func (t Time) Unix() int64 {
-	return time.Time(t).Unix()
-}
-
-func (t Time) UTC() time.Time {
-	return time.Time(t).UTC()
-}
-
-// Format returns t as a timestamp string with the provided layout.
-func (t Time) Format(layout string) string {
-	return time.Time(t).Format(layout)
-}
-
-// Time returns the JSON time as a time.Time instance in UTC
-func (t Time) Time() time.Time {
-	return time.Time(t).UTC()
-}
-
-// String returns t as a formatted string
-func (t Time) String() string {
-	return t.Time().String()
-}
-
 type Credentials struct {
 	Role                string
 	NoIpRestrict        bool
-	metaDataCredentials *AwsCredentials
+	metaDataCredentials *Credentials
 	MetadataRegion      string
-	LastRenewal         Time
+	LastRenewal         types.Time
 	mu                  sync.Mutex
 }
