--- pkg/internal/config/errors.go
@@ -1,27 +1,0 @@
-/*
- * Copyright 2020 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package config
-
-type Error string
-
-func (e Error) Error() string { return string(e) }
-
-const ClientCertificatesNotFoundError = Error("could not find client certificates")
-const EmbeddedConfigDisabledError = Error("embedded config is disabled")
-const HomeDirectoryError = Error("could not resolve user's home directory")
-const MissingTLSConfigError = Error("missing required mTLS configuration")
-const UnsupportedOSError = Error("running on unsupported operating system")
