--- pkg/internal/session/token.go
@@ -1,20 +1,0 @@
-package session
-
-import (
-	"github.com/netflix/weep/pkg/logging"
-)
-
-var sessions *tokenCache
-var log = logging.GetLogger()
-
-func init() {
-	sessions = createCache()
-}
-
-func GenerateToken(role string, ttlSeconds int) string {
-	return sessions.generateToken(role, ttlSeconds)
-}
-
-func CheckToken(token string) (bool, int) {
-	return sessions.checkToken(token)
-}
